Fixed Mosaicking. The issue had to due with the numpy array dimensions for the merged raster.
Because rasterio was merging the raster from a list of
raster images, all of which were being opened in numpy
arrays, the resulting merged numpy array had 3 dimensions
for it's shape instead of the normal two, with the
extra dimension relating to the number of images in
the list.

The issue was resolved using the numpy squeeze() function
to remove the extra dimension. This was then further refined
to exlude np.squeeze(), by specifying (arr.shape[1], arr.shape[2])
instead of (arr.shape[0], arr.shape[1]) in the rasterio
file writing code block at the end of tools.merge_chm().